[JLLGenerator] Initial draft of package for JLL generation
This provides a structured interface for generating JLL packages that
then rely upon JLLWrappers for compile-time code generation through
macros.  Note that this generates the new TOML-format JLLs (although it
does contain fields for code generation through on-load callbacks and
init blocks).  This package is yet somewhat incomplete, as it does not
encode some of the necessary but not functionally-required fields such
as license.