Properly fail builds, and provide ending environment
This adds `set -euo pipefail` to the top of all of our bash scripts,
which may be a little too strict, but honestly I kind of want to see how
far we can get with it.

This also fixes printing out the name of the command that failed, and
actually fails the build if something goes wrong.  A test is added to
ensure it stays that way.