Add build cache
This adds a build cache that remembers input content hashes (including
all dependencies and BB-related package treehashes) and skips the build
if the inputs have not changed since the last build.  It serializes a
mapping of input hashes to output artifact hashes (and associated
metadata such as build log, `env`, etc...).