Work out `--dry-run` and `extract_build_tarballs()` support
This adds a powerful new way of composing `build_tarballs.jl` scripts;
you can now "include" a `build_tarballs.jl` script via
`run_build_tarballs()`, and if you pass `--dry-run` as part of the
`ARGS` to that function, you can rescue out the `BuildConfig`,
`ExtractConfig` and `PackageConfig` structures for later building.  Even
more conveniently, you can now analyze a `PackageConfig` to determine if
it could have possibly originated from a `build_tarballs()` call, and if
so, regenerate the various arguments to `build_tarballs()` that would
result in that build.

This is being used to orchestrate some tests, and will be heavily
utilized in the bootstrapping scripts.