[Toolchains] Use `Libtool_jll` v`2.4.7+1`.
We need to ensure that we have the version that tries to use `/bin/bash`
instead of `/bin/sh`, as there are certain features that libtool
requires that are only available in a bourne-compatible shell.  The
current BinaryBuilder environment has `/bin/sh` as `bash`, so it
defaults to that in its built shebang lines, but that's not exactly
portable.