[Toolchains] Add `CMakeToolchain`
This ports over the `cmake` handling from BinaryBuilder.jl, but instead
of generating separate `cmake` toolchains for `clang` and `gcc`, we
simply create a single toolchain that uses `$CC`, `$LD`, etc...  This
greatly simplifies the implementation, making the assumption that in a
single build we'll really just want to use the `$CC` chosen by the
underlying `CToolchain`.  Through correct application of
`wrapper_prefixes` and `env_prefixes`, multiple `CMakeToolchain`s should
be able to coexist side-by-side in the same environment (e.g. a
`host-cmake` and a `target-cmake`).