Auditor: Use custom toolchain to provide `patchelf` and `install_name_tool`
Here we make an `AuditorToolchain` that allows us to grab properly
targeted binaries for modifying shared libraries/executables.
Previously we were just using `patchelf_jll` which is fine because any
`patchelf` can patch any ELF file, no matter the target.  However when I
added `CCTools_jll` in, I realized that I needed to select the
properly-targeted artifact for the current target platform, which is the
domain of our toolchains.