LazyJLLWrappers: Allow usage with non-lazy JLLs as dependencies
This works by simply pretending that there are no dynamic links between
the current Lazy JLL and the non-lazy dependency.  This is really not
recommended, but works for simple cases like running a Julia with a
Zlib_jll that is not lazy due to it being a stdlib.

The reason this works at all is that the non-lazy JLLs (by definition)
load all their libraries eagerly, which satisfies any Lazy JLLs that
depend on them, as long as the version constraints are reaslistically
satisfied.