Auditor: Add code paths and tests for macOS RPATH handling
Note that there are a few places we've needed to add `basename(soname)`
because macOS likes to embed the full path to a library.  Ideally, we
would instead look up the proper file it is pointing to, rather than
relying on our soname map, but this is not a very high priority at the
moment, because we do not expect to have conflicting SONAMEs in general.