Universe rework
This reworks our `Universe` to create a separate `BB2LocalRegistry`
where all JLL registrations are staged and clone `General` as a tarball
instead of as a git repository.  This significantly reduces the amount
of work needed to create a `Universe`, as well as simplifies universe
management when updating a persistent universe.