Toolchains:Fix `ar/ranlib/nm` with LTO, and `as` on macOS
Our `ar` and other tools needs to be pointed to `liblto_plugin.so` in
order for it to actually work on LTO objects.  Usually, this would be
handled by the Binutils install, but older Binutils versions don't
properly symlink to the actual install location from a global
`bfd_plugins` folder for some reason.  So we just manually pass in
`--plugin=...` for all versions.

The `as` on macOS also requires an `-arch` command to know which
assembler to invoke, we just add it in in the compiler wrapper.