JLLSource: Cache join resolution of artifacts per-prefix
Previously, we were caching the result of resolving each individual
`JLLSource`, but this ignores the fact that `collect_artifact_paths`
returns dependency JLL paths as well; causing situations where explicit
dependencies can go stale while implicit dependencies are re-resolved,
which results in deployment deploying two versions of the same
dependency.  We fix this by caching the result of the entire prefix,
which should be no less efficient (perhaps a little more so?) and more
correctly invalidated.