BinaryBuilderSources: Improve JLLSource caching
We need to change how the caching here works to cache both the
resolution of version constraints to gitshas as well as the artifact
paths themselves.  Practically speaking, this means that we need to
cache the `tree_hash` of the PackageSpec, and the `artifact_paths` of
the `JLLSource`.

I admit it is slightly inconvenient that we tie together Pkg resolution
and platform selection here; it means that we can't jointly resolve
artifacts for different platforms if we wanted to (although that would
be a rather unusual case).  That being said, I haven't run into a case
where that's absolutely necessary yet, so hopefully we never do.

The nasty case with the caching here has to do with things like Zlib_jll
not having a `tree_hash` after it's loaded from the cache, which is why
we need to make sure we mark down the `tree_hash` as well as the
artifact paths.