BinaryBuilderToolchains: Concretely type HostToolsToolchain.deps
We previously had this as an `AbstractSource[]`, but that doesn't buy us
anything as we're just using the `deps` to store the resolution result
of all our JLLs, it's the `toolchain_sources()` function that actually
defines the non-JLL dependencies (e.g. the `GeneratedSource`s and
whatnot).  Things are actually simplified by ensuring that everything
is a JLLSource here.