BuildTargetSpec: Alter environment variable names for default targets
When we define a "default" target via environment variables (so that
`$prefix` points to something, not just `$target_prefix` and
`$host_prefix`) we need to be careful here to not try and create a
mapping with an empty string as a key, as that corrupts the environment
block.  We alter the variable names here a bit so that the information
is still available within the build environment, but under legal,
predictable names.